@@ -1,5 +1,5 @@
 import 'package:flutter/foundation.dart';
-import 'package:flutter_riverpod/all.dart';
+import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 
 import '../init_get_it.dart';
@@ -7,7 +7,7 @@ import '../init_get_it.dart';
 final configProvider = ChangeNotifierProvider<Config>((_) => locator<Config>());
 
 class Config extends ChangeNotifier {
-  static SharedPreferences _sharedPrefs;
+  static late SharedPreferences _sharedPrefs;
 
   static const String _nameKey = "name";
   static const String _profileKey = "profile";
@@ -16,17 +16,25 @@ class Config extends ChangeNotifier {
     _sharedPrefs = await SharedPreferences.getInstance();
   }
 
-  set name(String value) {
+  set name(String? value) {
+    if (value == null) {
+      _sharedPrefs.remove(_nameKey);
+      return;
+    }
     _sharedPrefs.setString(_nameKey, value);
     notifyListeners();
   }
 
-  String get name => _sharedPrefs.getString(_nameKey);
+  String? get name => _sharedPrefs.getString(_nameKey);
 
-  set profile(String value) {
+  set profile(String? value) {
+    if (value == null) {
+      _sharedPrefs.remove(_profileKey);
+      return;
+    }
     _sharedPrefs.setString(_profileKey, value);
     notifyListeners();
   }
 
-  String get profile => _sharedPrefs.getString(_profileKey);
+  String? get profile => _sharedPrefs.getString(_profileKey);
 }
