@@ -9,7 +9,7 @@ import 'widgets/others.dart';
 import 'widgets/support_development.dart';
 
 class AboutScreen extends StatelessWidget {
-  const AboutScreen({Key key}) : super(key: key);
+  const AboutScreen({Key? key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
