@@ -4,10 +4,10 @@ import 'package:flutter/material.dart';
 
 class ImageBottomAppBar extends StatelessWidget {
   const ImageBottomAppBar({
-    Key key,
-    @required this.onTextPressed,
-    @required this.onDonePressed,
-    @required this.onColorPressed,
+    Key? key,
+    required this.onTextPressed,
+    required this.onDonePressed,
+    required this.onColorPressed,
   }) : super(key: key);
 
   final VoidCallback onColorPressed, onTextPressed, onDonePressed;
