@@ -9,7 +9,7 @@ import '../personalize_theme/personalize_theme.dart';
 import '../profile_screen/profile_screen.dart';
 
 class SettingsScreen extends StatelessWidget {
-  const SettingsScreen({Key key}) : super(key: key);
+  const SettingsScreen({Key? key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
