@@ -12,12 +12,12 @@ typedef BoolCallBack = bool Function();
 
 class WritingBottomAppBar extends StatefulWidget {
   const WritingBottomAppBar({
-    Key key,
-    @required this.onShareAsImage,
-    @required this.onShareAsText,
-    @required this.showSharePanel,
-    @required this.title,
-    @required this.poem,
+    Key? key,
+    required this.onShareAsImage,
+    required this.onShareAsText,
+    required this.showSharePanel,
+    required this.title,
+    required this.poem,
   }) : super(key: key);
 
   final VoidCallback onShareAsImage, onShareAsText;
@@ -30,7 +30,7 @@ class WritingBottomAppBar extends StatefulWidget {
 
 class _WritingBottomAppBarState extends State<WritingBottomAppBar>
     with SingleTickerProviderStateMixin {
-  AnimationController _iconController;
+  late AnimationController _iconController;
 
   @override
   void initState() {
@@ -44,7 +44,7 @@ class _WritingBottomAppBarState extends State<WritingBottomAppBar>
 
   @override
   void dispose() {
-    _iconController?.dispose();
+    _iconController.dispose();
     super.dispose();
   }
 
@@ -78,9 +78,10 @@ class _WritingBottomAppBarState extends State<WritingBottomAppBar>
                   AnimatedIconButton(
                     size: 25,
                     animationController: _iconController,
-                    startIcon: const Icon(Icons.share),
-                    endIcon: const Icon(Icons.close_rounded),
-                    onPressed: _changeIcon,
+                    icons: const [
+                      AnimatedIconItem(icon: Icon(Icons.share)),
+                      AnimatedIconItem(icon: Icon(Icons.close_rounded)),
+                    ],
                   ),
                 ],
               );
@@ -98,13 +99,5 @@ class _WritingBottomAppBarState extends State<WritingBottomAppBar>
         ],
       ),
     );
-  }
-
-  void _changeIcon() {
-    if (!widget.showSharePanel()) return;
-    if (_iconController.isCompleted)
-      _iconController.reverse();
-    else
-      _iconController.forward();
   }
 }
