@@ -1,8 +1,8 @@
-// This file is provided as a convenience for running integration tests via the
-// flutter drive command.
-//
-// flutter drive --driver integration_test/driver.dart --target integration_test/app_test.dart
+// // This file is provided as a convenience for running integration tests via the
+// // flutter drive command.
+// //
+// // flutter drive --driver integration_test/driver.dart --target integration_test/app_test.dart
 
-import 'package:integration_test/integration_test_driver.dart';
+// import 'package:integration_test/integration_test_driver.dart';
 
-Future<void> main() => integrationDriver();
+// Future<void> main() => integrationDriver();
